BigDecimal - Task, Report1

+++ src/main/java/pl/edu/agh/mwo/reporter/model/Task.java
@@ -1,15 +1,16 @@
 package pl.edu.agh.mwo.reporter.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Objects;
 
 public class Task {
     private final String name;
     private final LocalDate date;
-    private final int hours;
+    private final BigDecimal hours;
     private final String projectName;
 
-    public Task(String name, LocalDate date, int hours, String projectName) {
+    public Task(String name, LocalDate date, BigDecimal hours, String projectName) {
         this.name = name;
         this.date = date;
         this.hours = hours;
@@ -24,7 +25,7 @@ public class Task {
         return date;
     }
 
-    public int getHours() {
+    public BigDecimal getHours() {
         return hours;
     }
